refactor demarche select and demarches.json and add tests
