+++ src/pages/ApiParticulier/index.js
@@ -1,0 +1,201 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import Form from '../../components/Form';
+import Nav from '../../components/Nav';
+import OrganisationSection from '../../components/form-sections/OrganisationSection';
+import DemarcheSection from '../../components/form-sections/DemarcheSection';
+import DescriptionSection from '../../components/form-sections/DescriptionSection';
+import DonneesSection from '../../components/form-sections/DonneesSection';
+import CadreJuridiqueSection from '../../components/form-sections/CadreJuridiqueSection';
+import CguSection from '../../components/form-sections/CguSection';
+import DonneesPersonnellesSection from '../../components/form-sections/DonneesPersonnellesSection';
+import MiseEnOeuvreSection from '../../components/form-sections/MiseEnOeuvreSection';
+import demarches from './demarches.json';
+
+const DemarcheDescription = () => (
+  <div className="notification grey">
+    <p>
+      Pour avoir accès à l’API Particulier, diffusant des données personnelles,
+      vous devez obtenir un agrément. L’accès à cette API n’est pour l’instant
+      disponible que si vous êtes&nbsp;:
+    </p>
+    <ul>
+      <li>une administration</li>
+      <li>
+        une entreprise prestataire d’une administration ou ayant une délégation
+        de service public
+      </li>
+    </ul>
+    <p>
+      Pour utiliser API Particulier, vous devez vous engager à traiter la bonne
+      donnée par le bon agent de votre administration et informer correctement
+      l’usager.
+    </p>
+  </div>
+);
+
+const contacts = {
+  metier: {
+    heading: 'Contact métier',
+    description: () => (
+      <p>
+        Cette personne sera contactée en cas de problème fonctionnel sur votre
+        service.
+      </p>
+    ),
+    email: '',
+    phone_number: '',
+  },
+  technique: {
+    heading: 'Responsable technique',
+    description: () => (
+      <p>
+        Cette personne sera contactée en cas de problème technique sur votre
+        service. Le responsable technique peut être le contact technique de
+        votre prestataire.
+      </p>
+    ),
+    email: '',
+    phone_number: '',
+  },
+};
+
+const CadreJuridiqueDescription = () => (
+  <div className="text-quote">
+    <p>
+      Pour pouvoir bénéficier du raccordement à l&lsquo;API Particulier, le
+      cadre légal et réglementaire des fournisseurs de service doit permettre à
+      la DINUM de transmettre des données personnelles à votre entité
+      administrative.
+    </p>
+  </div>
+);
+
+const DonneesDescription = () => (
+  <div className="text-quote">
+    <p>
+      La loi informatique et libertés définit les principes à respecter lors de
+      la collecte, du traitement et de la conservation de données personnelles.
+    </p>
+    <p>L’article 6 précise :</p>
+    <ul>
+      <li>
+        3° [les données] sont adéquates, pertinentes et non excessives au regard
+        des finalités pour lesquelles elles sont collectées et de leurs
+        traitements ultérieurs ;
+      </li>
+      <li>
+        4° Elles sont exactes, complètes et, si nécessaire, mises à jour ; les
+        mesures appropriées doivent être prises pour que les données inexactes
+        ou incomplètes au regard des finalités pour lesquelles elles sont
+        collectées ou traitées soient effacées ou rectifiées ;
+      </li>
+    </ul>
+    <p>
+      Nous vous remercions de sélectionner uniquement les données strictement
+      nécessaires à votre téléservice. Le non-respect du principe de
+      proportionnalité vous expose vis à vis de la CNIL.
+    </p>
+  </div>
+);
+
+const availableScopes = [
+  {
+    value: 'dgfip_avis_imposition',
+    label: "DGFIP - Avis d'imposition",
+  },
+  {
+    value: 'dgfip_adresse',
+    label: 'DGFIP - Adresse',
+  },
+  {
+    value: 'cnaf_quotient_familial',
+    label: 'CNAF - Quotient familial',
+  },
+  {
+    value: 'cnaf_allocataires',
+    label: 'CNAF - Allocataires',
+  },
+  {
+    value: 'cnaf_enfants',
+    label: 'CNAF - Enfants',
+  },
+  {
+    value: 'cnaf_adresse',
+    label: 'CNAF - Adresse',
+  },
+];
+
+const ApiParticulier = ({
+  match: {
+    params: { enrollmentId },
+  },
+}) => (
+  <div className="dashboard">
+    <Nav
+      navLinks={[
+        { id: 'head', label: 'Formulaire', style: { fontWeight: 'bold' } },
+        { id: 'organisation', label: 'Organisation' },
+        { id: 'service-numerique', label: 'Service numérique' },
+        { id: 'description', label: 'Description' },
+        { id: 'donnees', label: 'Données' },
+        { id: 'cadre-juridique', label: 'Cadre juridique' },
+        { id: 'donnees-personnelles', label: 'Données personnelles' },
+        { id: 'contacts-moe', label: 'Mise en œuvre' },
+        { id: 'cgu', label: "Modalités d'utilisation" },
+      ]}
+      contactInformation={[
+        {
+          email: 'contact@particulier.api.gouv.fr',
+          label: 'Nous contacter',
+          subject: 'Contact%20via%20datapass.api.gouv.fr',
+        },
+      ]}
+    />
+    <div className="main">
+      <Form
+        enrollmentId={enrollmentId}
+        target_api="api_particulier"
+        title="Demande d'accès à API Particulier"
+        DemarcheDescription={DemarcheDescription}
+      >
+        <OrganisationSection />
+        <DemarcheSection demarches={demarches} />
+        <DescriptionSection
+          intitulePlaceholder={
+            '« Calcul du quotient familial pour la facturation scolaire et périscolaire »'
+          }
+        />
+        <DonneesSection
+          availableScopes={availableScopes}
+          DonneesDescription={DonneesDescription}
+        />
+        <CadreJuridiqueSection
+          CadreJuridiqueDescription={CadreJuridiqueDescription}
+        />
+        <DonneesPersonnellesSection />
+        <MiseEnOeuvreSection initialContacts={contacts} />
+        <CguSection cguLink="https://particulier.api.gouv.fr/API_Particulier_modalites.pdf" />
+      </Form>
+    </div>
+  </div>
+);
+
+ApiParticulier.propTypes = {
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      enrollmentId: PropTypes.string,
+    }),
+  }),
+};
+
+ApiParticulier.defaultProps = {
+  match: {
+    params: {
+      enrollmentId: null,
+    },
+  },
+};
+
+export default ApiParticulier;
